refactor: UNT-T23078: Restructured Ripple Button In Spinner Button
