--- src/components/RippleButton/index.ts
@@ -1,1 +1,0 @@
-export { default as RippleButton } from './RippleButton';
