+++ src/components/index.ts
@@ -5,4 +5,3 @@ export * from './LinearGradient';
 export * from './RadialGradient';
 export * from './SpinnerButton';
 export * from './SpinnerComponent';
-export * from './Button';
