+++ src/styles/SpinnerButtonStyle.ts
@@ -4,6 +4,12 @@ export default StyleSheet.create({
   buttonContainer: {
     flex: 1,
   },
+  button: {
+    alignItems: 'center',
+    justifyContent: 'center',
+    height: 50,
+    width: '100%',
+  },
   centerAlign: {
     justifyContent: 'center',
     alignItems: 'center',
@@ -17,4 +23,17 @@ export default StyleSheet.create({
   defaultGradientContainerStyle: {
     width: '100%',
   },
+  animatedViewContainer: {
+    marginTop: 0,
+    marginBottom: 0,
+    marginLeft: 0,
+    marginRight: 0,
+    paddingTop: 0,
+    paddingBottom: 0,
+    paddingLeft: 0,
+    paddingRight: 0,
+  },
+  rippleButtonStyle: {
+    width: '100%',
+  }
 });
