+++ src/utils/index.tsx
@@ -1,5 +1,6 @@
 export { default as getPercentageFromAngle } from './getPercentageFromAngle';
 export { default as useAnimatedValues } from './useAnimatedValues';
+export { default as useRippleButton } from './useRippleButton';
 export {
   DEFAULT_COLOR_WHITE,
   DEFAULT_ANIMATION_TYPE,
